Add QueryDSL dependencies and extract custom methods from GameRankingRepository
