+++ server/build.gradle
@@ -30,5 +30,7 @@ dependencies {
 	compile('org.springframework.boot:spring-boot-starter-web')
 	compile('com.allanditzel:spring-security-csrf-token-filter:1.1')
 	compile('com.google.guava:guava:21.0')
+	compile("com.querydsl:querydsl-jpa:4.1.4")
+	compile("com.querydsl:querydsl-apt:4.1.4:jpa")
 	testCompile('org.springframework.boot:spring-boot-starter-test')
 }
