+++ server/src/main/java/tabletop/repositories/game/GameRankingRepositoryCustom.java
@@ -1,0 +1,14 @@
+package tabletop.repositories.game;
+
+import tabletop.domain.game.Game;
+import tabletop.domain.ranking.GameRanking;
+import tabletop.domain.user.User;
+
+import java.util.List;
+
+public interface GameRankingRepositoryCustom {
+
+    void updateGameRanking(Game game, List<User> usersByResult);
+
+    List<GameRanking> getRankingForGame(List<User> users, Game game);
+}
