+++ server/src/main/java/tabletop/services/GameRankingService.java
@@ -3,9 +3,9 @@ package tabletop.services;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import tabletop.domain.game.Game;
-import tabletop.domain.ranking.TournamentRanking;
+import tabletop.domain.ranking.GameRanking;
 import tabletop.domain.user.User;
-import tabletop.repositories.GameRankingRepository;
+import tabletop.repositories.game.GameRankingRepository;
 
 import java.util.List;
 
@@ -14,7 +14,7 @@ public class GameRankingService {
     @Autowired
     private GameRankingRepository gameRankingRepository;
 
-    public List<TournamentRanking> getRankingForGame(List<User> users, Game game) {
+    public List<GameRanking> getRankingForGame(List<User> users, Game game) {
         return gameRankingRepository.getRankingForGame(users, game);
     }
